+++ protected/humhub/modules/user/components/PeopleQuery.php
@@ -104,42 +104,48 @@ class PeopleQuery extends ActiveQueryUser
             ->where(['directory_filter' => 1])
             ->andWhere(['IN', 'internal_name', array_keys($fields)])
             ->all();
-        $checkedFilteredFields = [];
+
+        $conditions = [];
         foreach ($filteredProfileFields as $filteredField) {
             /* @var $filteredField ProfileField */
-            if (!isset($fields[$filteredField->internal_name])) {
+            $fieldName = $filteredField->internal_name;
+            if (!isset($fields[$fieldName])) {
                 // Skip unknown field
                 continue;
             }
 
+            $value = $fields[$fieldName];
             $fieldType = $filteredField->getFieldType();
+
             if ($fieldType instanceof CheckboxList) {
-                $condition = 'REGEXP';
+                $delimiter = preg_quote(CheckboxList::MULTI_VALUE_DELIMITER, '/');
+                $condition = ['OR',
+                    // Search if match a line, i.e. a key of the available options of the Checkbox List,
+                    // in the multi value string separated with CheckboxList::MULTI_VALUE_DELIMITER
+                    ['REGEXP', 'profile.' . $fieldName, '(^|' . $delimiter . ')' . preg_quote($value, '/') . '(' . $delimiter . '|$)'],
+                ];
+                if ($fieldType->allowOther) {
+                    // Search in "Other:" field, only relevant while the option is enabled -
+                    // if disabled, the column may still hold stale values that must not match.
+                    $condition[] = ['=', 'profile.' . CheckboxList::getOtherColumnName($fieldName), $value];
+                }
             } elseif ($fieldType instanceof Select) {
-                $condition = '=';
+                $condition = ['=', 'profile.' . $fieldName, $value];
             } else {
-                $condition = 'LIKE';
+                $condition = ['LIKE', 'profile.' . $fieldName, $value];
             }
 
-            $checkedFilteredFields[$filteredField->internal_name] = [
-                'value' => $fields[$filteredField->internal_name],
-                'condition' => $condition,
-            ];
+            $conditions[] = $condition;
         }
 
-        if (empty($checkedFilteredFields)) {
+        if (empty($conditions)) {
             return $this;
         }
 
         $this->joinWith('profile');
 
-        foreach ($checkedFilteredFields as $field => $data) {
-            $condition = $data['condition'];
-            $value = $data['value'];
-            $search = ($condition === 'REGEXP')
-                ? '(^|\\n)' . preg_quote($value, '/') . '(\\n|$)' // Search if match a line, i.e. a key of the available options of the Checkbox List
-                : $value;
-            $this->andWhere([$condition, 'profile.' . $field, $search]);
+        foreach ($conditions as $condition) {
+            $this->andWhere($condition);
         }
 
         return $this;